renamed reactors to better names
