+++ src/prerna/auth/utils/reactors/admin/AdminGetEngineUsageReactor.java
@@ -1,36 +1,37 @@
-package prerna.engine.impl.model.inferencetracking;
+package prerna.auth.utils.reactors.admin;
 
 import java.util.List;
 import java.util.Map;
 
 import prerna.auth.User;
-import prerna.auth.utils.SecurityEngineUtils;
+import prerna.auth.utils.SecurityAdminUtils;
 import prerna.auth.utils.SecurityQueryUtils;
+import prerna.engine.impl.model.inferencetracking.ModelInferenceLogsUtils;
 import prerna.reactor.AbstractReactor;
 import prerna.sablecc2.om.PixelDataType;
 import prerna.sablecc2.om.ReactorKeysEnum;
 import prerna.sablecc2.om.nounmeta.NounMetadata;
 
-public class GetOverAllEngineUsageReactor extends AbstractReactor {
+public class AdminGetEngineUsageReactor extends AbstractReactor {
+
 	
-	public GetOverAllEngineUsageReactor() {
+	public AdminGetEngineUsageReactor() {
 		this.keysToGet = new String[]{ReactorKeysEnum.ENGINE.getKey(), ReactorKeysEnum.LIMIT.getKey(), ReactorKeysEnum.OFFSET.getKey(),};
 	}
-
+	
 	@Override
 	public NounMetadata execute() {
 		User user = this.insight.getUser();
+		SecurityAdminUtils adminUtils = SecurityAdminUtils.getInstance(user);
+		if(adminUtils == null) {
+			throw new IllegalArgumentException("User must be an admin to perform this function");
+		}
 		organizeKeys();
 		String engineId = this.keyValue.get(this.keysToGet[0]);
 		if(engineId == null || engineId.isEmpty()) {
 			throw new IllegalArgumentException("Must input an engine id");
 		}
-		engineId = SecurityQueryUtils.testUserEngineIdForAlias(user, engineId);
-		if(!SecurityEngineUtils.userIsOwner(user, engineId)) {
-			throw new IllegalArgumentException("Engine does not exist or user is not an owner of Engine");
-		}
-		
-		
+		engineId = SecurityQueryUtils.testUserEngineIdForAlias(this.insight.getUser(), engineId);
 		String limit = this.keyValue.get(this.keysToGet[1]);
 		String offset = this.keyValue.get(this.keysToGet[2]);
 		String dateFilter = this.keyValue.get(ReactorKeysEnum.DATE_FILTER.getKey());
